pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oilpriceapi"
version = "1.0.0"
description = "Official Python SDK for OilPriceAPI - Real-time and historical oil prices"
authors = [
    {name = "OilPriceAPI", email = "support@oilpriceapi.com"}
]
maintainers = [
    {name = "OilPriceAPI Team", email = "support@oilpriceapi.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
keywords = [
    "oil", "prices", "api", "commodities", "energy",
    "brent", "wti", "natural gas", "trading", "finance"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.5.0; python_version < '3.10'",
]

[project.optional-dependencies]
pandas = [
    "pandas>=1.5.0",
    "numpy>=1.20.0",
]
async = [
    "aiohttp>=3.8.0",
]
cache = [
    "redis>=4.5.0",
    "cachetools>=5.3.0",
]
cli = [
    "click>=8.0.0",
    "rich>=13.0.0",
]
all = [
    "oilpriceapi[pandas,async,cache,cli]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.261",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://oilpriceapi.com"
Documentation = "https://docs.oilpriceapi.com/sdk/python"
Repository = "https://github.com/oilpriceapi/python-sdk"
Issues = "https://github.com/oilpriceapi/python-sdk/issues"
Changelog = "https://github.com/oilpriceapi/python-sdk/blob/main/CHANGELOG.md"

[project.scripts]
oilprice = "oilpriceapi.cli:main"

[tool.setuptools]
packages = ["oilpriceapi"]

[tool.setuptools.package-data]
oilpriceapi = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []
fixable = ["ALL"]

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true
pretty = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["."]
addopts = [
    "--cov=oilpriceapi",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v",
]

[tool.coverage.run]
source = ["oilpriceapi"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"